move util fns
