+++ console/src/components/Services/EventTrigger/StreamingLogs/Logs.js
@@ -7,7 +7,7 @@ import Tabs from 'react-bootstrap/lib/Tabs';
 import Tab from 'react-bootstrap/lib/Tab';
 import RedeliverEvent from '../TableCommon/RedeliverEvent';
 import TableHeader from '../TableCommon/TableHeader';
-import parseRowData from './util';
+import { parseRowData } from '../utils';
 import {
   loadEventLogs,
   setTrigger,
@@ -295,32 +295,32 @@ class StreamingLogs extends Component {
                 >
                   {finalResponse.status_code
                     ? [
-                        'Status Code: ',
-                        verifySuccessStatus(finalResponse.status_code) ? (
-                          <i
-                            className={
-                              styles.invocationSuccess + ' fa fa-check'
-                            }
-                          />
-                        ) : (
-                          <i
-                            className={
-                              styles.invocationFailure + ' fa fa-times'
-                            }
-                          />
-                        ),
-                        finalResponse.status_code,
-                        ' ',
-                        <OverlayTrigger
-                          placement="top"
-                          overlay={tooltip.statusCodeDescription}
-                        >
-                          <i
-                            className="fa fa-question-circle"
-                            aria-hidden="true"
-                          />
-                        </OverlayTrigger>,
-                      ]
+                      'Status Code: ',
+                      verifySuccessStatus(finalResponse.status_code) ? (
+                        <i
+                          className={
+                            styles.invocationSuccess + ' fa fa-check'
+                          }
+                        />
+                      ) : (
+                        <i
+                          className={
+                            styles.invocationFailure + ' fa fa-times'
+                          }
+                        />
+                      ),
+                      finalResponse.status_code,
+                      ' ',
+                      <OverlayTrigger
+                        placement="top"
+                        overlay={tooltip.statusCodeDescription}
+                      >
+                        <i
+                          className="fa fa-question-circle"
+                          aria-hidden="true"
+                        />
+                      </OverlayTrigger>,
+                    ]
                     : null}
                 </div>
                 <AceEditor
